% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparation_Expand_TableDescription.R
\name{addEmptyRowsBeforeNewResource}
\alias{addEmptyRowsBeforeNewResource}
\title{Add Empty Rows Before New Resource Entries in a Data Table}
\usage{
addEmptyRowsBeforeNewResource(table)
}
\arguments{
\item{table}{A \code{data.table} object with a column named \code{resource}. The function
will insert empty rows based on the conditions specified.}
}
\value{
A modified \code{data.table} with empty rows inserted before new resource entries.
}
\description{
This function inserts an empty row in a data table before each new resource entry,
except before the first row and where the first column (\code{resource}) is not NA.
It's useful for visually separating different resources in a data table.
}
\examples{
\dontrun{
# Assuming `table` is a data.table with a column `resource`
table <- data.table(resource = c(NA, 'Resource1', 'Resource1', NA, 'Resource2'),
                    value = 1:5)
table <- addEmptyRowsBeforeNewResource(table)
print(table)
}
}
