% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparation_Expand_TableDescription.R
\name{checkResult}
\alias{checkResult}
\title{Check the Result of Expanded Table Description for Constraints}
\usage{
checkResult(expanded_table_description)
}
\arguments{
\item{expanded_table_description}{A data.table object containing the expanded table description.
This table must include at least the columns 'column_name' and 'resource'.}
}
\value{
Logical value indicating whether the expanded table description passes the checks.
Returns \code{TRUE} if all checks pass (i.e., no column names exceed the maximum length and there are no duplicated column names),
otherwise \code{FALSE}.
}
\description{
Evaluates the expanded table description for specific constraints, such as the maximum length of column names
allowed in Postgres databases, and checks for duplicated column names. It prints error messages for violations
and provides suggestions for resolution.
}
\details{
The function performs two main checks:
\enumerate{
\item It checks if any column names exceed 64 characters, which is the maximum allowed length for column names in Postgres databases.
If any column names are too long, it prints an error message with the offending column names and suggests defining a replacement
in the 'table_description_collapsed' section of the Table_Description_Definition.xlsx file to shorten these names.
\item It checks for duplicated column names within each 'resource' group. If duplicates are found, it prints an error message
with the duplicated names and suggests checking the 'fhir_expression' entries in the Table_Description_Definition.xlsx file
for potential duplicates.
}

Error messages include specific solutions and notes to help address the identified issues.
}
\examples{
# Assuming `expanded_table_description` is a data.table object obtained from expanding a table description:
# result <- checkResult(expanded_table_description)
# if (!result) {
#  message("There were errors in the expanded table description.")
# } else {
#  message("The expanded table description passed all checks.")
# }

}
