% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_Convert_Column_Types.R
\name{convertType}
\alias{convertType}
\title{Convert columns of specific types in tables}
\usage{
convertType(resource_tables, convert_columns, convert_type_function)
}
\arguments{
\item{resource_tables}{A named list of data.table objects to be written to the database.
Each element in the list represents a dataset intended for a corresponding table in the database.}

\item{convert_columns}{A character vector specifying the types of columns to be converted.}

\item{convert_type_function}{A function used to convert the columns.}
}
\description{
This function converts columns of specific types in tables using a given conversion function.
}
\details{
This function takes a list of tables, where each table is represented by a data frame,
and converts columns of specific types specified by \code{convert_columns}. The conversion is
performed using the provided \code{convert_type_function}. The function assumes that there is
a table description table containing columns: "resource", "column_name", and "type".
}
