% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_Convert_Column_Types.R
\name{convertTypes}
\alias{convertTypes}
\title{Convert Types}
\usage{
convertTypes(resource_tables)
}
\arguments{
\item{resource_tables}{A list of data tables representing resources.}
}
\value{
This function modifies the input resource tables by converting the data types of columns.
}
\description{
This function converts data types of columns in the provided resource tables.
}
\details{
This function takes a list of data tables \code{resource_tables} as input.
It converts the data types of columns in these tables based on predefined mappings:
\itemize{
\item Columns marked as "datetime" are converted to POSIXct objects using \code{etlutils::convertDateTimeFormat}.
\item Columns marked as "date" are converted to date objects using \code{etlutils::convertDateFormat}.
\item Columns marked as "time" are converted to POSIXct objects representing time of the day using \code{etlutils::convertTimeFormat}.
\item Columns marked as "int" are converted to integer values using \code{etlutils::convertIntegerFormat}.
\item Columns marked as "decimal" are converted to numeric (floating-point) values using \code{etlutils::convertDecimalFormat}.
\item Columns marked as "boolean" are converted to logical (TRUE/FALSE) values using \code{etlutils::convertBooleanFormat}.
}
}
