% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_Load_Resources.R
\name{createWardPatitentIDPerDateTable}
\alias{createWardPatitentIDPerDateTable}
\title{Create a data.table with ward and patient ID per date.}
\usage{
createWardPatitentIDPerDateTable(patientIDsPerWard)
}
\arguments{
\item{patientIDsPerWard}{A list of patient IDs, where each element corresponds to a ward.}
}
\value{
A data.table with columns date_time, ward, and pid, representing the date, ward,
and patient ID for each combination extracted from the provided list.
}
\description{
This function takes a list of patient IDs per ward and constructs a data.table
with columns for date_time, ward, and pid. Each row represents a unique combination
of date, ward, and patient ID extracted from the provided list.
}
\examples{
library(data.table)
# Example: A list of patient IDs per ward
patientIDsPerWard <- list(
  Ward_A = c("PID_A001", "PID_A002", "PID_A003"),
  Ward_B = c("PID_B001", "PID_B002"),
  Ward_C = c("PID_C001", "PID_C002", "PID_C003", "PID_C004")
)

# Applying the function
result_table <- createWardPatitentIDPerDateTable(patientIDsPerWard)

# Displaying the result
print(result_table)

}
