% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparation_Expand_TableDescription.R
\name{expandTableDescriptionFromFile}
\alias{expandTableDescriptionFromFile}
\title{Expand Table Description from an Excel File}
\usage{
expandTableDescriptionFromFile(
  table_description_collapsed_excel_simple_filename
)
}
\arguments{
\item{table_description_collapsed_excel_simple_filename}{The filename of the Excel file containing
the table description and expansion tables. The filename should not include the path or '.xlsx' extension.}
}
\value{
The function does not explicitly return a value but performs the expansion of the table description
as a side effect, utilizing other functions within the package to read the Excel file, extract relevant tables,
and expand the table description according to predefined rules.
}
\description{
This function reads a table description from an Excel file and expands it using specified expansion tables
also contained within the same Excel file. It is designed to facilitate the process of table expansion
by directly working with file-based inputs. The function automatically appends a '.xlsx' extension if not present,
determines the correct file path based on the execution context (interactive or package), reads the Excel file
into a list of tables, extracts the collapsed table description, and then applies the expansion logic.
}
\examples{
# Assuming the Excel file 'Table_Description_Definition.xlsx' exists in the appropriate directory
# and contains the necessary table description and expansion tables:
# expandTableDescriptionFromFile("Table_Description_Definition")

}
\seealso{
\code{\link[etlutils]{readExcelFileAsTableList}}, \code{\link{expandTableDescription}}
}
