% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparation_Expand_TableDescription.R
\name{extractReplacePatterns}
\alias{extractReplacePatterns}
\title{Extract Replacement Patterns from a Table}
\usage{
extractReplacePatterns(table_description_collapsed)
}
\arguments{
\item{table_description_collapsed}{A data table that contains the patterns and their replacements.
The table must have at least 'resource' and 'resource_prefix' columns.}
}
\value{
A named list where each name-value pair corresponds to a pattern and its replacement string.
}
\description{
Extracts replacement patterns and their corresponding strings from a given table. This function
is designed to parse a table where specific rows indicate the start of pattern-replacement pairs.
Patterns are expected to be in the 'resource' column, and their corresponding replacement strings
in the 'resource_prefix' column. The function uses a helper function to find the first row that
contains the headers for patterns and replacements, and then iterates over the rows to collect
these pairs into a list.
}
\examples{
library(data.table)
# Implementing the function with simulated data
table_description_collapsed <- data.table(
  resource = c("Line 1", "pattern", "pattern1", "pattern2"),
  resource_prefix = c("Line 1", "replacement", "replace1", "replace2")
)
# Apply the function
result <- extractReplacePatterns(table_description_collapsed)
# Print the result
print(result)

}
\seealso{
\code{\link[etlutils]{getFirstRowWithPatterns}}, \code{\link[etlutils]{isSimpleNotEmptyString}}
}
