% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Get_Table_Descriptions.R
\name{getTableDescriptionsTable}
\alias{getTableDescriptionsTable}
\title{Get a table with the Table Descriptions for all fhir resources from Excel file}
\usage{
getTableDescriptionsTable(columns = NA)
}
\arguments{
\item{columns}{column names which should be in the return table. If NA then no column
will be removed.}
}
\value{
A data table with the table descriptions.
}
\description{
This function reads the "Table_Description.xlsx" file from the "extdata" directory
in the package and returns the data as a data frame.
}
