% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kds2db-package.R
\docType{package}
\name{kds2db-package}
\alias{kds2db}
\alias{kds2db-package}
\title{kds2db: ETL process to extract all FHIR data for specific Patients to a database}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
