% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_Load_Resources.R
\name{loadReferencedResourcesByOwnIDFromFHIRServer}
\alias{loadReferencedResourcesByOwnIDFromFHIRServer}
\title{Load Referenced Resources by Own ID from FHIR Server}
\usage{
loadReferencedResourcesByOwnIDFromFHIRServer(
  table_descriptions,
  resource_tables
)
}
\arguments{
\item{table_descriptions}{A list containing the fhircrackr table descriptions for the result
tables, including \code{reference_types} which lists all resources that reference other resources.}

\item{resource_tables}{A list of data.tables, each containing FHIR resources for specific
patients. This list is updated with the referenced resources.}
}
\value{
An updated list of data.tables including the referenced resources.
}
\description{
This function loads FHIR resources referenced by other resources from a FHIR server. It takes
a set of table descriptions and a list of resource tables, identifies all unique resource names
in the \code{reference_types} column of the table descriptions, and loads the referenced resources
for each unique resource name. The result is an updated list of resource tables including the
newly loaded referenced resources.
}
