% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_Load_Resources.R
\name{loadResourcesByPatientIDFromFHIRServer}
\alias{loadResourcesByPatientIDFromFHIRServer}
\title{Load FHIR resources for a given set of patient IDs and create a table of ward-patient ID per date.}
\usage{
loadResourcesByPatientIDFromFHIRServer(
  patient_IDs_per_ward,
  table_descriptions
)
}
\arguments{
\item{patient_IDs_per_ward}{A list of patient IDs, where each element corresponds to a ward.}

\item{table_descriptions}{the fhircrackr table descriptions of the result tables}
}
\value{
A list of data.tables, each containing FHIR resources for a specific patient,
and the last element is a table representing the ward and patient ID per date.
}
\description{
This function takes a list of patient IDs per ward, extracts unique patient IDs,
loads FHIR resources for those patient IDs from the FHIR server using the provided
\code{TABLE_DESCRIPTION}, and creates an additional table of ward-patient ID per date. The
result is a list of data.tables, where each element contains FHIR resources for a specific
patient, and the last element is a table representing the ward and patient ID per date.
}
