% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_Load_Resources.R
\name{loadResourcesFromFHIRServer}
\alias{loadResourcesFromFHIRServer}
\title{Load Resources and Referenced Resources from FHIR Server}
\usage{
loadResourcesFromFHIRServer(patient_IDs_per_ward, table_descriptions)
}
\arguments{
\item{patient_IDs_per_ward}{A list of patient IDs, where each element corresponds to a ward and
contains patient IDs associated with that ward.}

\item{table_descriptions}{A list containing two elements: \code{pid_dependant} and
\code{pid_independant}, each of which describes table structures for resources that are dependent
and independent of patient IDs, respectively.}
}
\description{
This function loads resources for a given set of patient IDs per ward from a FHIR server and
then loads any additional referenced resources. It uses two steps: first, loading resources
directly associated with patient IDs using \code{loadResourcesByPatientIDFromFHIRServer}, and second,
loading resources referenced by the initially loaded resources using
\code{loadReferencedResourcesByOwnIDFromFHIRServer}. The results are then saved as RData files, with
filenames derived from the resource names.
}
\details{
The function iterates through all resources loaded in both steps and saves them as
RData files using \code{polar_write_rdata}. The filenames are derived by converting the names of the
resources in the \code{resource_tables} list to lowercase.
}
