% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Get_Table_Descriptions.R
\name{loadTableDescriptionFile}
\alias{loadTableDescriptionFile}
\title{Load Table Description Excel File}
\usage{
loadTableDescriptionFile()
}
\value{
A data table with table descriptions
}
\description{
This function loads a table description excel file
}
