% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_Main.R
\name{retrieve}
\alias{retrieve}
\title{Starts the retrieval for this project. This is the main start function start the ETL job
from FHIR to Database}
\usage{
retrieve()
}
\description{
Starts the retrieval for this project. This is the main start function start the ETL job
from FHIR to Database
}
